Added some new checks to avoid exceptions thrown to screen/other things when deleting users and/or activity types
git-svn-id: http://agilefant.googlecode.com/svn/trunk@432 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/web/UserAction.java
@@ -29,6 +29,16 @@ public class UserAction extends ActionSupport implements CRUDAction{
 	}
 
 	public String delete() {
+		User u = userDAO.get(userId);
+		if (u == null){
+			super.addActionError(super.getText("user.notFound"));
+			return Action.ERROR;
+		}
+		if(u.getAssignables().size() > 0 || u.getWatchedBacklogItems().size() > 0 
+				|| u.getWatchedTasks().size() > 0) {
+			super.addActionError(super.getText("user.hasLinkedItems"));
+			return Action.ERROR;
+		}
 		userDAO.remove(userId);
 		return Action.SUCCESS;
 	}
